requirements.txt
streamlit>=1.34.0
pyairtable>=2.1.0
pandas>=2.0.0
plotly>=5.17.0
python-dotenv>=1.0.0
supabase>=2.0.0
matplotlib
